modules/api/src/main/java/io/spring/microservice/api/CustomerController.java
package io.spring.microservice.api;

import java.util.List;

import javax.validation.Valid;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import io.spring.microservice.models.CustomerDto;
import io.spring.microservice.persistence.services.CustomerService;

@RestController
@RequestMapping(path = "/api/v1/customers")
public class CustomerController {

    private final CustomerService customerService;

    @Autowired
    public CustomerController(CustomerService customerService) {
        this.customerService = customerService;
    }

    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
    public List<CustomerDto> getAllCustomers() {
        return customerService.getAllCustomers();
    }

    @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
    @ResponseStatus(HttpStatus.CREATED)
    public void insertCustomer(@Valid @RequestBody CustomerDto customer) {
        customerService.saveCustomer(customer);
    }

    @DeleteMapping
    @RequestMapping(path = "/deleteById/{id}")
    public void deleteCustomerById(@PathVariable Long id) {
        customerService.deleteCustomer(id);
    }

    @DeleteMapping
    @RequestMapping(path = "/deleteByLastName/{name}")
    public void deleteCustomerByLastName(@PathVariable(name = "name") String lastName) {
        customerService.deleteCustomerByLastName(lastName);
    }

}
